Fix format specifier for MAINPID in sd_notifyf

+++ SystemdConnector/SystemdConnector.cpp
@@ -53,8 +53,8 @@ namespace Plugin {
         int rc = sd_notifyf(0,
             "READY=1\n"
             "STATUS=Platform Server is Ready (from WPE Framework Compositor Plugin)\n"
-            "MAINPID=%d",
-            ::getpid());
+            "MAINPID=%ld",
+            static_cast<long>(::getpid()));
         if (rc) {
             TRACE(Trace::Error, (_T("Notify Nexus Server Ready to systemd: FAILED (%d)\n"), rc));
         } else {
